Demo
Tic tac toe is a two-player game wherein the players will have to choose either X or O in a 3×3 grid. 
The first player can choose any position in the grid and both will get one move at a time, one after the other. The player who succeeds in placing three of their marks in a horizontal, vertical, or diagonal row wins the game. 
In the backend, the game will maintain a 2-D array to save the current state of the 3×3 grid. A function will be made to check if the box that’s clicked by the player in the displaying grid is empty or not and will decide to put up an image of O if the previous one was X and vice versa. And when the consecutive images match, the game ends. 

Tools/ languages – Android Studio, Java, XML.